Splitting configurations into multiple files. AKA: The longer the PR is open, the more stuff I will add to it.
